audio-player/requirements.txt
udi_interface>=3.0.57
python-vlc

